+++ tests/repository/elements.py
@@ -1,4 +1,7 @@
+from typing import List
+
 from protean.core.aggregate import BaseAggregate
+from protean.core.repository.base import BaseRepository
 from protean.core.field.basic import Integer, String
 
 
@@ -6,3 +9,8 @@ class Person(BaseAggregate):
     first_name = String(max_length=50, required=True)
     last_name = String(max_length=50, required=True)
     age = Integer(default=21)
+
+
+class PersonRepository(BaseRepository):
+    def find_adults(self, age: int = 21) -> List[Person]:
+        pass
